+++ packages/obsidian/CHANGELOG.md
@@ -3,6 +3,108 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.20](https://github.com/rowthan/pagenote/compare/@pagenote/obsidian@0.1.15...@pagenote/obsidian@0.1.20) (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/obsidian
+
+
+
+
+
+## 0.1.20-next.6 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/obsidian
+
+
+
+
+
+## 0.1.20-next.5 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/obsidian
+
+
+
+
+
+## 0.1.20-next.4 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/obsidian
+
+
+
+
+
+## 0.1.20-next.3 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/obsidian
+
+
+
+
+
+## 0.1.20-next.2 (2024-01-07)
+
+
+### Features
+
+* **shared:** remove $links ([18a77b1](https://github.com/rowthan/pagenote/commit/18a77b11fbe1ffcede483f7af12ed0506c2afee9))
+
+
+
+
+
+## 0.1.20-next.1 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/obsidian
+
+
+
+
+
+## 0.1.20-next.0 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/obsidian
+
+
+
+
+
+## 0.1.19-next.0 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/obsidian
+
+
+
+
+
+## 0.1.16 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/obsidian
+
+
+
+
+
+## 0.1.15 (2024-01-07)
+
+
+### Features
+
+* auto focus after create ([d17e52d](https://github.com/rowthan/pagenote/commit/d17e52d7279791415aebb5b58c8c792baa1a5606))
+
+
+
+
+
+## 0.1.14 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/obsidian
+
+
+
+
+
 ## 0.1.12 (2024-01-06)