+++ packages/shared/CHANGELOG.md
@@ -3,6 +3,92 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.14.29](https://github.com/rowthan/pagenote/compare/@pagenote/shared@1.14.24...@pagenote/shared@1.14.29) (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/shared
+
+
+
+
+
+## 1.14.29-next.4 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/shared
+
+
+
+
+
+## 1.14.29-next.3 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/shared
+
+
+
+
+
+## 1.14.29-next.2 (2024-01-07)
+
+
+### Features
+
+* **shared:** remove $links ([18a77b1](https://github.com/rowthan/pagenote/commit/18a77b11fbe1ffcede483f7af12ed0506c2afee9))
+
+
+
+
+
+## 1.14.29-next.1 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/shared
+
+
+
+
+
+## 1.14.29-next.0 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/shared
+
+
+
+
+
+## 1.14.28-next.0 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/shared
+
+
+
+
+
+## 1.14.25 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/shared
+
+
+
+
+
+## 1.14.24 (2024-01-07)
+
+
+### Features
+
+* auto focus after create ([d17e52d](https://github.com/rowthan/pagenote/commit/d17e52d7279791415aebb5b58c8c792baa1a5606))
+
+
+
+
+
+## 1.14.23 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/shared
+
+
+
+
+
 ## 1.14.22 (2024-01-06)