+++ packages/sync/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.1.16-next.0 (2024-01-07)
+
+**Note:** Version bump only for package @pagenote/sync
+
+
+
+
+
 ## 0.1.15-next.0 (2023-12-31)